Refactors on the UI package structure.
